Move net::tcp tests to their own files
